fromproto5+fromproto6: Inline logic from Parameter() function in ArgumentsData() function
