--- internal/fwfunction/xfwfunction/doc.go
@@ -1,6 +1,0 @@
-// Copyright (c) HashiCorp, Inc.
-// SPDX-License-Identifier: MPL-2.0
-
-// Package xfwfunction contains additional functions for use with provider defined functions.
-// This package is separate from the fwfunction package to prevent import cycles.
-package xfwfunction
